backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.0

# AI / Embeddings
anthropic==0.42.0
openai==1.58.1
tiktoken==0.8.0

# Supabase
supabase==2.11.0
postgrest==0.19.1

# Document parsing
python-docx==1.1.2
pypdf2==3.0.1

# HTTP
httpx==0.28.1
python-multipart==0.0.20

# Database (direct Postgres for migrations)
psycopg2-binary==2.9.11
